requirements.txt
pandas>=1.1.4
pathlib>=1.0.1
python-docx>=0.8.10
selenium>=3.141.0
bs4>=0.0.1
urllib3>=1.26.2

